+++ src/lib/actions/create.mjs
@@ -10,7 +10,6 @@ import { determineBucketName } from '../shared/determine-bucket-name'
 import { getCredentials } from './lib/get-credentials'
 import { getSiteTag } from '../shared/get-site-tag'
 import * as plugins from '../plugins'
-import { progressLogger } from '../shared/progress-logger'
 import { SiteTemplate } from '../shared/site-template'
 import { syncSiteContent } from './lib/sync-site-content'
 import { trackStackStatus } from './lib/track-stack-status'
@@ -61,42 +60,6 @@ const create = async ({
   } else {
     return { success, stackName }
   }
-}
-
-const createCertificate = async ({ acmClient, apexDomain }) => {
-  progressLogger.write(`Creating wildcard certificate for '${apexDomain}'...`)
-  const input = { // RequestCertificateRequest
-    DomainName              : '*.' + apexDomain, // TODO: support more narrow cert?
-    ValidationMethod        : 'DNS', // TODO: support email
-    SubjectAlternativeNames : [
-      apexDomain, 'www.' + apexDomain
-    ], /*
-    // IdempotencyToken: "STRING_VALUE", TODO: should we use this?
-    /* DomainValidationOptions: [ // DomainValidationOptionList : TODO: is this only used for email verification?
-      { // DomainValidationOption
-        DomainName: "STRING_VALUE", // required
-        ValidationDomain: "STRING_VALUE", // required
-      },
-    ], */
-    Options : { // CertificateOptions
-      CertificateTransparencyLoggingPreference : 'ENABLED'
-    },
-    // CertificateAuthorityArn: "STRING_VALUE", TODO: only used for private certs, I think
-    /* Tags: [ // TagList : TODO: support tags? tag with the website
-      { // Tag
-        Key: "STRING_VALUE", // required
-        Value: "STRING_VALUE",
-      },
-    ], */
-    KeyAlgorithm : 'RSA_2048' // TODO: support key options"RSA_1024" || "RSA_2048" || "RSA_3072" || "RSA_4096" || "EC_prime256v1" || "EC_secp384r1" || "EC_secp521r1",
-  }
-  // this method can safely be called multiple times; it'll  match  existing certs (by domain name I'd assume)
-  const command = new RequestCertificateCommand(input)
-  const response = await acmClient.send(command)
-
-  const { CertificateArn } = response
-
-  return CertificateArn
 }
 
 const createSiteStack = async ({ credentials, noDeleteOnFailure, siteInfo }) => {
